ensure blocksmith will NOT cause white screen crash
